// src/types/ApiKeyRequest.ts
// Code generated by schema2code at 2026-03-07 19:34:27. DO NOT EDIT.


/**
 * Request to create an API key for IDE and external service integration
 */
export interface ApiKeyRequest {
  /**
   * Human-readable name for this key
   */
  name: string;
  /**
   * Number of days until the key expires (null = never)
   */
  expires_in_days?: number;
  /**
   * API scopes this key has access to (e.g., 'chat', 'generate', 'embed')
   */
  scopes: (string)[];
}